requirements.txt
annotated-types     
anyio               
APScheduler         
bcrypt              
blinker             
certifi             
cffi                
charset-normalizer  
click               
cryptography        
Deprecated          
fastapi             
Flask               
Flask-Limiter       
h11                 
httptools           
idna                
importlib_resources 
itsdangerous        
Jinja2              
limits              
markdown-it-py      
MarkupSafe          
mdurl               
ordered-set         
packaging           
pip                 
psycopg2            
pycparser           
pydantic            
pydantic_core       
Pygments            
PyJWT               
python-dotenv       
python-multipart    
pytz                
PyYAML              
requests            
rich                
setuptools          
simpleeval          
six                 
slowapi             
sniffio             
starlette           
typing_extensions   
tzlocal             
urllib3             
uvicorn             
uvloop              
watchfiles          
websockets          
Werkzeug            
wrapt               